feat: movie test data setup
feat: 영화 테스트 데이터 setup 구현

+++ server/src/main/java/com/ticketing/server/movie/domain/Movie.java
@@ -1,7 +1,6 @@
 package com.ticketing.server.movie.domain;
 
 import com.ticketing.server.global.dto.repository.AbstractEntity;
-import com.ticketing.server.movie.service.dto.MovieDto;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.validation.constraints.NotNull;
@@ -15,11 +14,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Movie extends AbstractEntity {
 
-    @NotNull
-    @Column(unique = true)
-    private String title;
+	@NotNull
+	@Column(unique = true)
+	private String title;
 
-    @NotNull
-    private Integer runningTime;
+	@NotNull
+	private Long runningTime;
 
 }
